V2.0 Implementacion de buscador

+++ app/src/main/java/com/skysam/hchirinos/transport/ui/bookings/BookingFragment.kt
@@ -1,28 +1,29 @@
 package com.skysam.hchirinos.transport.ui.bookings
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.widget.SearchView
+import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.Lifecycle
 import androidx.recyclerview.widget.RecyclerView
 import com.skysam.hchirinos.transport.R
 import com.skysam.hchirinos.transport.dataClasses.Booking
-import com.skysam.hchirinos.transport.dataClasses.Bus
 import com.skysam.hchirinos.transport.databinding.FragmentBookingBinding
 import com.skysam.hchirinos.transport.ui.common.WrapLayoutManager
 import com.skysam.hchirinos.transport.ui.payment.AddPaymentDialog
 import com.skysam.hchirinos.transport.ui.payment.ViewDetailsDialog
 
-class BookingFragment : Fragment(), OnClick {
+class BookingFragment : Fragment(), OnClick, MenuProvider, SearchView.OnQueryTextListener {
 
     private var _binding: FragmentBookingBinding? = null
     private val binding get() = _binding!!
     private val viewModel: BookingViewModel by activityViewModels()
     private lateinit var bookingAdapter: BookingAdapter
     private val bookings = mutableListOf<Booking>()
+    private val listSearch = mutableListOf<Booking>()
     private lateinit var wrapLayoutManager: WrapLayoutManager
 
     override fun onCreateView(
@@ -31,6 +32,7 @@ class BookingFragment : Fragment(), OnClick {
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentBookingBinding.inflate(inflater, container, false)
+        binding.topAppBar.addMenuProvider(this, viewLifecycleOwner, Lifecycle.State.RESUMED)
         return binding.root
     }
 
@@ -60,19 +62,41 @@ class BookingFragment : Fragment(), OnClick {
                     binding.tvListEmpty.visibility = View.VISIBLE
                     binding.rvBookings.visibility = View.GONE
                 } else {
-                    val listTemp = mutableListOf<Booking>()
-                    for (item in bookings) {
-                        listTemp.add(item)
+                    if (listSearch.isEmpty()) {
+                        val listTemp = mutableListOf<Booking>()
+                        for (item in bookings) {
+                            listTemp.add(item)
+                        }
+                        bookings.clear()
+                        bookings.addAll(it)
+
+                        if (listTemp.size == bookings.size) updateItem(listTemp)
+                        if (listTemp.size < bookings.size) addItem(listTemp)
+                        if (listTemp.size > bookings.size) deleteItem(listTemp)
+
+                        binding.tvListEmpty.visibility = View.GONE
+                        binding.rvBookings.visibility = View.VISIBLE
+                    } else {
+                        bookings.clear()
+                        bookings.addAll(it)
+                        for (book in it) {
+                            for (bookSearch in listSearch) {
+                                if (book.id == bookSearch.id && book != bookSearch) {
+                                    listSearch[listSearch.indexOf(bookSearch)] = book
+                                }
+                            }
+                        }
+                        var exists = true
+                        var bookingRemove: Booking? = null
+                        for (bookSearch in listSearch) {
+                            if (!it.contains(bookSearch)) {
+                                exists = false
+                                bookingRemove = bookSearch
+                            }
+                        }
+                        if (!exists) listSearch.remove(bookingRemove)
+                        bookingAdapter.updateList(listSearch)
                     }
-                    bookings.clear()
-                    bookings.addAll(it)
-
-                    if (listTemp.size == bookings.size) updateItem(listTemp)
-                    if (listTemp.size < bookings.size) addItem(listTemp)
-                    if (listTemp.size > bookings.size) deleteItem(listTemp)
-
-                    binding.tvListEmpty.visibility = View.GONE
-                    binding.rvBookings.visibility = View.VISIBLE
                 }
             }
         }
@@ -138,5 +162,49 @@ class BookingFragment : Fragment(), OnClick {
         viewModel.viewBooking(booking)
         val addPaymentDialog = AddPaymentDialog()
         addPaymentDialog.show(requireActivity().supportFragmentManager, tag)
+    }
+
+    override fun onCreateMenu(menu: Menu, inflater: MenuInflater) {
+        val itemSearch = menu.findItem(R.id.app_bar_search)
+        val search = itemSearch.actionView as SearchView
+        search.isSubmitButtonEnabled = true
+        search.setOnQueryTextListener(this)
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+        return when(menuItem.itemId) {
+            R.id.app_bar_search -> true
+            else -> false
+        }
+    }
+
+    override fun onQueryTextSubmit(query: String?): Boolean {
+        return false
+    }
+
+    override fun onQueryTextChange(newText: String?): Boolean {
+        if (bookings.isNotEmpty()) {
+            val userInput: String = newText!!.lowercase()
+            listSearch.clear()
+
+            if (userInput.isNotEmpty()) {
+                for (booking in bookings) {
+                    if (booking.name.lowercase().contains(userInput)) {
+                        listSearch.add(booking)
+                    }
+                }
+                if (listSearch.isEmpty()) {
+                    binding.lottieAnimationView.visibility = View.VISIBLE
+                    binding.lottieAnimationView.playAnimation()
+                } else {
+                    binding.lottieAnimationView.visibility = View.GONE
+                }
+                bookingAdapter.updateList(listSearch)
+            } else {
+                bookingAdapter.updateList(bookings)
+                binding.lottieAnimationView.visibility = View.GONE
+            }
+        }
+        return true
     }
 }
